Add robust crash bar detection
Add contour detection to the color masking algorithm to obtain a bounding box of the crash bar without relying on a hard-coded rectangular region. Also switch color masking from sRGB to HSV to increase robustness.

Switched difference calculation to normalized cross-correlation, which has mixed results. Keeping for now, but we may want to revert this to the original loss function.